requirements.txt
functions-framework==3.*
python-dotenv
tweepy
pillow